refactor: move Airbyte package to internal
